// assets/scripts/Enum.ts
export enum RulesType{
    NineBall = 'nineball',
    ThreeCushion = 'threecushion',
    Snooker = 'snooker',
    FourteenOne = 'fourteenone'
}

export enum BallState{
    Stationary = "Stationary",
    Rolling = "Rolling",
    Sliding = "Sliding",
    Falling = "Falling",
    InPocket = "InPocket",
}